#5: Move game info and plugin info to configs folder
